requirements.txt
-i https://pypi.org/simple
flask==1.0.2
numpy==1.17.4
tensorflow==1.14.0
opencv-python-headless
werkzeug==0.16.0
matplotlib==3.1.1
gunicorn==19.0.0
Pillow==6.0.0
image
scipy